feat: show permit balance in card when no favorite action is visible
When a permit is selected and neither the add-favorite nor remove-favorite
button is shown, display the current balance (icon + value) in that spot.

- Backend: expose balance_unit and remaining_minutes (float) in the
  WebSocket status payload so the card can render the balance without an
  extra round-trip
- Backend: fix float truncation — remaining_balance was cast to int,
  losing decimals for monetary (EURO) permits
- Frontend: render a ha-formfield balance chip with icon and label,
  styled and aligned identically to the add/remove favorite controls
- Frontend: format balance per unit type — MINUTE → Xh Ym,
  TIMES → integer count, monetary → currency symbol + amount (€12.50)

+++ custom_components/city_visitor_parking/models.py
@@ -61,7 +61,7 @@ class CoordinatorData:
     """Coordinator data for entities and services."""
 
     permit_id: str
-    permit_remaining_minutes: int
+    permit_remaining_minutes: float
     permit_balance_unit: str | None
     zone_validity: tuple[TimeRange, ...]
     reservations: tuple[Reservation, ...]
